# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Documentation",
        "description": "Admin OpenApi description",
        "contact": {
            "email": "matrizlab@info.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Admin API Server"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Collection"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Create"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User Delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "e351648aebb146847ec4190628eabbf4",
                "responses": {
                    "200": {
                        "description": "Authenticated User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/info": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "f2af1c2554ee72ba65b7dfa8fccacd13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Info Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "34e093df10270da0e999d0bd07f70995",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Password Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateInfoRequest": {
                "title": "Update User Info request",
                "description": "Update User Info request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdatePasswordRequest": {
                "title": "Update User Password request",
                "description": "Update User Password request body data",
                "properties": {
                    "password": {
                        "title": "password",
                        "type": "string"
                    },
                    "password_confirm": {
                        "title": "password_confirm",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "title": "Store User request",
                "description": "Store User request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "title": "Update User request",
                "description": "Update User request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}